requirements.txt

h5py==3.1.0
numpy==1.18.1
Pillow==7.0.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.42.1
zipp==3.4.0
